Updates models to allow FOGI parameterization in implicit models.
Moves ExplicitOpModel._add_reparameterization to OpModel base class
and generalizes (not much needed) so that this method works for all
model types.  Pushes ImplicitOpModel.errorgen_coefficients method into
LocalNoiseModel and CloudNoiseModel derived classes, as the implementation
needs to be different in the two cases.
